chore(templates): add default impls for normally unused functions
arcgis doesn't complain if you leave these out, but i figure it makes more sense to include
them with an empty impl that returns the right thing than to leave them out.
it's unclear what arcgis does when it fails to call these functions when they're missing.

they're included in a block that ignores formatting so that they can be as minimal as possible in the
source file, as they are just noise.